Added nullChecker and implented null test in forms

+++ CVFeedbackApp/CVFeedbackApp/SetOptionsForm.cs
@@ -16,67 +16,101 @@ namespace CVFeedbackApp
         {
             InitializeComponent();
         }
+        private bool GetErrorResult()
+        {
+            if (ErrorChecker.nullTextboxErrorCheck("Option Title", OptionTitleTextBox.Text)
+                && ErrorChecker.nullTextboxErrorCheck("Option Message", OptionMessageTextBox.Text))
+            {
+                return true;
+            }
+            else
+                return false;
+        }
         private void OptionAddition()
         {
-            //Gets/Creates intance of optionSet and Option
-            OptionSet OptionSetInstance = OptionSet.GetOptionSet();
-            Option OptionInstance = Option.GetOption();
+            if (GetErrorResult())
+            {
+                //Gets/Creates intance of optionSet and Option
+                OptionSet OptionSetInstance = OptionSet.GetOptionSet();
+                Option OptionInstance = Option.GetOption();
 
-            //populates option from textboxes
-            OptionInstance.SetOptionTitle(OptionTitleTextBox.Text);
-            OptionInstance.SetOptionMessage(OptionMessageTextBox.Text);
+                //populates option from textboxes
+                OptionInstance.SetOptionTitle(OptionTitleTextBox.Text);
+                OptionInstance.SetOptionMessage(OptionMessageTextBox.Text);
 
-            //Populates OptionSet with new Option
-            OptionSetInstance.AddOptionToSet(OptionInstance);
+                //Populates OptionSet with new Option
+                OptionSetInstance.AddOptionToSet(OptionInstance);
 
-            //Sets Both Instances
-            Option.SetOption(OptionInstance);
-            OptionSet.SetNewOptionSet(OptionSetInstance);
+                //Sets Both Instances
+                Option.SetOption(OptionInstance);
+                OptionSet.SetNewOptionSet(OptionSetInstance);
+            }
+            else
+            {
+                //Does nothing
+            }
+            
 
 
         }
         private void AddOptionToSet_MouseClick(object sender, MouseEventArgs e)
         {
-            this.Hide();
+            if(GetErrorResult())
+            {
+                this.Hide();
 
-            //Calls Option addition method
+                //Calls Option addition method
 
-            OptionAddition();
-
-
-            //Deletes Option for new option to be added
-            Option.DeleteOptionInstance();
+                OptionAddition();
 
 
-            //creates new form (loop)
-            SetOptionsForm OptionForm2 = new SetOptionsForm();
-            OptionForm2.ShowDialog(); 
+                //Deletes Option for new option to be added
+                Option.DeleteOptionInstance();
+
+
+                //creates new form (loop)
+                SetOptionsForm OptionForm2 = new SetOptionsForm();
+                OptionForm2.ShowDialog();
+            }
+            else
+            {
+                //Does Nothing
+            }
+            
         }
 
         private void NewOptionSet_MouseClick(object sender, MouseEventArgs e)
         {
-            this.Hide();
+            if (GetErrorResult())
+            {
+                this.Hide();
 
-            //Calls Option addition method
+                //Calls Option addition method
 
-            OptionAddition();
+                OptionAddition();
 
-            //Gets intance of Generic template and OptionSet
-            GenericTemplate genericTemplateInstance = GenericTemplate.GetGenericTemplate();
-            OptionSet optionSetInstance = OptionSet.GetOptionSet();
+                //Gets intance of Generic template and OptionSet
+                GenericTemplate genericTemplateInstance = GenericTemplate.GetGenericTemplate();
+                OptionSet optionSetInstance = OptionSet.GetOptionSet();
 
-            //Feeds OptionSet to Generic Template Instance
+                //Feeds OptionSet to Generic Template Instance
 
-            genericTemplateInstance.AddOptionSet(optionSetInstance);
+                genericTemplateInstance.AddOptionSet(optionSetInstance);
 
-            //Deletes Current OptionSet
+                //Deletes Current OptionSet
 
-            OptionSet.DeleteOptionSet();
+                OptionSet.DeleteOptionSet();
 
-            //Creates and displays AddOptionSetForm
+                //Creates and displays AddOptionSetForm
 
-            AddOptionSet Optionset2 = new AddOptionSet();
-            Optionset2.ShowDialog();
+                AddOptionSet Optionset2 = new AddOptionSet();
+                Optionset2.ShowDialog();
+            }
+            else
+            {
+                //Does Nothing
+            }
+            
         }
 
 
@@ -87,30 +121,37 @@ namespace CVFeedbackApp
         /// <param name="e"></param>
         private void FinishTemplate_MouseClick(object sender, MouseEventArgs e)
         {
-            //Creates instance Of Generic Template
-            GenericTemplate genericTemplateInstance = GenericTemplate.GetGenericTemplate();
-
-            //Displays Message form asking for confirmation
-            MessageBoxButtons buttons = MessageBoxButtons.YesNo;
-            DialogResult textboxInput;
-            textboxInput = MessageBox.Show("Do you want to save this changes into " + genericTemplateInstance.GetTitle() + "?", "Save Progress", buttons);
-            if ( textboxInput == DialogResult.Yes )
+            if (GetErrorResult())
             {
-                //creates instance of DBConnection and calls insertion method
-                DBConnection connectionInstance = DBConnection.getDatabaseConection();
-                connectionInstance.insertToDB(null, genericTemplateInstance);
+                //Creates instance Of Generic Template
+                GenericTemplate genericTemplateInstance = GenericTemplate.GetGenericTemplate();
 
-                //Deletes GenericTemplate
-                GenericTemplate.DeleteGenericTemplateInstance();
+                //Displays Message form asking for confirmation
+                MessageBoxButtons buttons = MessageBoxButtons.YesNo;
+                DialogResult textboxInput;
+                textboxInput = MessageBox.Show("Do you want to save this changes into " + genericTemplateInstance.GetTitle() + "?", "Save Progress", buttons);
+                if (textboxInput == DialogResult.Yes)
+                {
+                    //creates instance of DBConnection and calls insertion method
+                    DBConnection connectionInstance = DBConnection.getDatabaseConection();
+                    connectionInstance.insertToDB(null, genericTemplateInstance);
+
+                    //Deletes GenericTemplate
+                    GenericTemplate.DeleteGenericTemplateInstance();
 
 
-                this.Hide();
-                FirstMenuForm newFirstMenuForm = new FirstMenuForm();
-                newFirstMenuForm.ShowDialog();
+                    this.Hide();
+                    FirstMenuForm newFirstMenuForm = new FirstMenuForm();
+                    newFirstMenuForm.ShowDialog();
+                }
+                else
+                {
+                    //Does nothing
+                }
             }
             else
             {
-                //does nothing
+                //Does Nothing
             }
